Forward messages from Slack to IRC
This commit's a mess, I have some cleanup to do for later!

We now have an HTTP server listening for Slack events,
we query slack to learn about the channels and users,
and when we receive a message we forward it to the right IRC channel.

Note that the IRC server currently doesn't show anyone but the
connected client in the members list of each channel.

As for performance, well, this cobbled together piece of hack
is designed to handle maybe 5 concurrent users before it starts
wildly timing out on the Slack event acknowledgments.
Doing O(n) network requests synchronously when you're supposed
to ack an event immediately is Not Smart.
But hey, it's fairly easily fixable Later™.